Issue #108: Mark ApiGateway certificates as CA certs
Rebased with master and integrated the change for issue #108
which allows upgrading of client auth certificates to CA
certificates. This is a workaround for the AWS API GW client
auth certificates which don't have the flags set to work with Go.

The configuration for 1.2 differs that the aws.apigw.cert.cn
property was dropped in favor of a caupgcn parameter for the
newly introduced certificate sources.