Move arg parsing code to flagset
Add KVVar(), KVSliceVar() and StringSliceVar() methods
in favor of custom code for parsing args. The approach still
has some caveats:

proxy.addr does not conform to the new format since
it does not start with 'key=' but just the 'host:port' value.
This is partially an oversight and partially required for the
backwards compatibility with the legacy listener. Maybe this needs
to move to a different variable like 'proxy.listen' to clean this up
properly.

The second problem is that KV slice is a comma-sep list of semicolon
delimited k=v pairs. This prevents 'v' to contain commas since parsing
becomes ambgious. There either needs to be a separate delimiter - which
then conflicts with the string slice - or the comma needs to be escaped.