Issue #134: Refactor vault token support
The initial vault support implementation required a token
that had root or sudo privileges since fabio would create a
renewable orphan token to be indpendent of the originals
token lifecycle and to allow providing fabio with a
short-lived token for bootstrapping.

Requiring root and/or sudo privileges is providing fabio
with more privileges than it needs. Also, the decoupling
from the parent token also means that revoking the parent
will not revoke fabios access to vault.

This patch changes this behavior so that fabio will no
longer try to create a token of its own but use the provided
token. To prevent sniffing the token should be wrapped
(requires vault >= 0.6.0). fabio will transparently unwrap
the token.

The provided token can be an orphan and should be renewable
for the duration fabio is expected to run. It is recommended
not to set the explicit_max_ttl unless fabio is restarted
before that time expires.

fabio needs the following policies set on the path where the
certificates are stored.

  path "secret/fabio/cert" {
    capabilities = ["list"]
  }

  path "secret/fabio/cert/*" {
    capabilities = ["read"]
  }