Issue #1: Add TCP proxy with SNI support
This patch adds support for configuring a listener
to provide a passthrough TCP proxy for TLS connections
and use the SNI server extension in the ClientHello
message for routing. Targets need to register their
prefix as 'host/' for now.

This implementation is EXPERIMENTAL since it is functional
but not ready for production use since it does not yet set
proper timeouts on connections. It also does not have a full
integration test. It is merged to gather more real-world
test data to complete the implementation.