Issue #172: Consul cert store URL with token not parsed correctly
A consul certificate store is configured with a URL which points to a
key in the KV store which stores the certificates. This URL can be
secured with a consul auth token which is appended to the URL with
?token=xxx. When parsing the key name from the URL the token and other
parameters are not removed producing an invalid key in the KV store.
This patch fixes this and strips off the token and other parameters.