Issue #199: Refactor config loader test
The config loader test was not testing the full
roundtrip of the command line arguments, environment
variables and config files from files and URLs.
Various smaller tests were only testing individual parts
of the config loader. This made updating and refactoring
the config loading code more vulnerable to regressions.

This refactor replaces all smaller tests with a single
test suite that tests all aspects of the config loading
in the same way as it is used by the main() function.
New tests can be added easily and consistently.

At the same time, variables which store temporary
values for the config loader have been removed
from the main config structure as they are
redundant and confusing since no part of the
application needs access to the raw argument
values.