Issue #186: runtime error: integer divide by zero
When fabio is presented with more than 200 targets for a single
route then a rounding error had the effect that no target would
get traffic which subsequently lead to a division by zero error
in a code path which did not expect an empty list of targets.

This change modifies the distribution algorithm as follows:

* The case where all targets receive an equal amount of traffic
has now a fast path which just returns the list of targets.

* For the other case a larger ring of 10.000 slots is always
used to achieve a certain amount of accuracy. In addition, all
targets that have a non-zero weight will use at least one slot
to work around rounding issues and the fact that some servers
would not receive traffic although they are alive and healthy.

This should guarantee an unlimited number of instances for a
single route.