Refactor listener and server code
Create a common function to create a listening socket which supports TLS
from a CertSource and the PROXY protocol which can then be used for all
TCP based servers.

Pronounce the difference between listeners which hold the socket for
incoming connections, servers which manage connections and configuration
and handlers which handle individual requests.

Introduce a separate TCP server which is similar to the HTTP server and
unify the ListenAndServe code across both servers. This allows using
both the TCP and HTTP server to be used for (test-)servers and proxy
servers depending on the handler they use.

Drop the custom graceful shutdown logic in favor of the context based
shutdown introduced with Go 1.8. This makes Go 1.8 the minimum
requirement.

Drop the legacy and unused websocket handler which uses
golang.org/x/net/websocket.

Move the server and listening code into the proxy package. The HTTP
related code remains under the proxy package instead of an http package
to avoid naming conflicts with the standard http library. Instead, the
files are prefixed with http_.