Added -registry.consul.register.tlsskipverify flag
When serving the UI/API over HTTPS, the fabio consul
health check will most likely fail as it's trying to
connect to the page over its IP. Most certificates don't
include IP SANs. This flag allows users to toggle whether
or not to skip TLS verification for this particular check.