Issue #279: fabio does not start with multiple listeners
This patch fixes a bug in the startServers() function which
spawns the listeners and the proxy servers. Since the
listeners are started in go routines the loop var with the
listener config needs to be captured either by passing it to
the go routine as a parameter or by creating a local value
inside the loop. Otherwise, only the last value is
preserved.

Fixes #279