Issue #280: Extend Forwarded header with protocol data
This patch adds an 'httpproto' field which contains the lower case
version of the HTTP protocol to the Forwarded header of all upstream
requests.

When fabio terminates the TLS connection two additional fields 'tlsver'
and 'tlscipher' are added to the Forwarded field. 'tlsver' contains the
TLS version as 'tlscipher' the chosen cipher suite. The cipher suite is
formatted as a hex number from
https://golang.org/pkg/crypto/tls/#pkg-constants.

Fixes #280