Add option to register service alias for fabio
This patch adds a "register=<name>" option which tells Fabio to register
a new frontend service in Consul on behalf of the requesting backend
service.  The new service provides Consul-hosted A records to Fabio with
the desired service name, provided as the <name> argument to "register".

Fabio will deregister front end services when the backend services
leave, and will deregister all Fabio-registered services on exit.