WIP: add circuit breaker to detect network failures
This patch adds a circuit breaker monitor which issues route table
fragments to delete routes for which the circuit breaker is open.

The current implementation supports only the HTTP path. TCP and WS will
follow.

The integration test isn't testing the behavior automatically yet
but can be used to eyeball the behavior. This needs to be improved.

Also missing is the code to GC the circuit breakers that aren't
tripped over time but that should be easy to implement in this
code.

Fixes #393