Issue #408: log consul state changes as DEBUG
This patch changes the log level for Consul raft log changes from
INFO to DEBUG so that most users will not see this anymore. High
change rate is usually an indicator for flapping services or health
checks with volatile output like timestamps but that problem should be
dealt with on the Consul side.

A PR for #368 will to add an indicator for the change rate of the Consul
state.

Fixes #408