Issue #460: Fix access logging when gzip is enabled
The code that captured the content length and the status code
for the access logger depended on the http.Handler to be of a
certain type. Wrapping the handler in a gzip handler broke this
silently.

This fix removes that dependency and instead wraps the response
writer directly to capture the values required for access logging.

Fixes #460