Issue #421: Close websocket connection on failure
The websocket proxy is implemented as a raw tcp proxy
which relies on the client and server to close the
connection. When a websocket upgrade fails the upstream
server may keep the connection open.

If a proxy like nginx is used in front of fabio
it will keep its connection to fabio open effectively
establishing a direct channel between nginx and the
upstream server which will be used for any request
forwarded by nginx to fabio.

Adding a 'Connection: close' header to the upstream
request should indicate to the server to close the
connection. If that works then we can keep the raw
tcp proxy for websockets. Otherwise, fabio needs
to handle the handshake and close the connection
itself.

Fixes #421