Make VaultSource compatible with KV Backend V2
Detect if a KV backend uses the new v2 versioning features and rewrite
request paths and bodies if necessary.

The new API uses additional /data/ and /metadata/ for GET/PUT and LIST
operations, respectively. To facilitate versioning, v2 wraps the actual
payload in a JSON object with "data" and "metadata" keys:

    {
      "data": {<payload (same as v1)>},
      "metadata": { <versioning info> }
    }