docs: fix layout without JS enabled
The Bootstrap stylesheet for the website sets the header to position:
fixed, taking it out of the layout flow. Some javascript later updates
the style attribute, so that the content is pushed down by an
appropriate amount.

If JS is disabled that correction doesn't happen, and so the header
overlaps the content.

Make the header static by default, and only apply position: fixed if
JavaScript is enabled.

Fixes #599

Signed-off-by: Peter Schultz <peter.schultz@classmarkets.com>