Deprecate deregisterCriticalServiceAfter option
The deregisterCriticalServiceAfter option applied to the HTTP health
check for services that are registered in Consul.

Commit ea3a36552b introduced additional Consul TTL checks. These TTL
checks have DeregisterCriticalServiceAfter set to the minimum of one
minute and will therefore always beat or tie the HTTP check's
DeregisterCriticalServiceAfter option, rendering it obsolete.

Document this fact in the usage message and HTML docs.

Signed-off-by: Peter Schultz <peter.schultz@classmarkets.com>